refactor(ng-packagr): remove sass worker and fallback to pure-JS sass
Removes the legacy Piscina-based Sass worker thread pool implementation (`sass-worker-implementation.ts`, `worker.ts`, and `worker-pool.ts`). When the native `sass-embedded` compiler cannot be used (such as inside WebContainers or when disabled), the build system now directly falls back to the in-process pure-JS Dart Sass asynchronous compiler (`sass.initAsyncCompiler()`).

Consolidates the Sass compilation logic in `src/lib/styles/sass` into a simplified `SassCompiler` class in `sass-service.ts`.

### Environment Variable
- Replaces `NG_BUILD_SASS_WORKER` with `NG_BUILD_SASS_EMBEDDED`.
- When `NG_BUILD_SASS_EMBEDDED` is set to `0` or `false`, or when running within a WebContainer environment (`process.versions.webcontainer`), the native embedded Sass compiler is disabled and the pure-JS Sass compiler is used instead.
