+++ package.json
@@ -45,7 +45,6 @@
     "jsonc-parser": "^3.3.1",
     "less": "^4.2.0",
     "ora": "^9.0.0",
-    "piscina": "^5.0.0",
     "postcss": "^8.4.47",
     "rolldown": "^1.2.0",
     "rolldown-plugin-dts": "^0.28.0",
