+++ src/lib/styles/sass/rebasing-importer.ts
@@ -30,7 +30,9 @@ abstract class UrlRebasingImporter implements Importer<'sync'> {
    * @param rebaseSourceMaps When provided, rebased files will have an intermediate sourcemap added to the Map
    * which can be used to generate a final sourcemap that contains original sources.
    */
-  constructor(private entryDirectory: string) {}
+  constructor(private entryDirectory: string) {
+    this.load = this.load.bind(this);
+  }
 
   abstract canonicalize(url: string, options: CanonicalizeContext): URL | null;
 
@@ -111,6 +113,7 @@ export class RelativeUrlRebasingImporter extends UrlRebasingImporter {
     private directoryCache = new Map<string, DirectoryEntry>(),
   ) {
     super(entryDirectory);
+    this.canonicalize = this.canonicalize.bind(this);
   }
 
   canonicalize(url: string, options: CanonicalizeContext): URL | null {
@@ -286,32 +289,6 @@ export class RelativeUrlRebasingImporter extends UrlRebasingImporter {
 }
 
 /**
- * Provides the Sass importer logic to resolve module (npm package) stylesheet imports via both import and
- * use rules and also rebase any `url()` function usage within those stylesheets. The rebasing will ensure that
- * the URLs in the output of the Sass compiler reflect the final filesystem location of the output CSS file.
- */
-export class ModuleUrlRebasingImporter extends RelativeUrlRebasingImporter {
-  constructor(
-    entryDirectory: string,
-    directoryCache: Map<string, DirectoryEntry>,
-    private finder: (specifier: string, options: CanonicalizeContext) => URL | null,
-  ) {
-    super(entryDirectory, directoryCache);
-  }
-
-  override canonicalize(url: string, options: CanonicalizeContext): URL | null {
-    if (url.startsWith('file://')) {
-      return super.canonicalize(url, options);
-    }
-
-    let result = this.finder(url, options);
-    result &&= super.canonicalize(result.href, options);
-
-    return result;
-  }
-}
-
-/**
  * Provides the Sass importer logic to resolve module (npm package) stylesheet imports asynchronously
  * and also rebase any `url()` function usage within those stylesheets.
  */
@@ -324,6 +301,8 @@ export class AsyncModuleUrlRebasingImporter implements Importer<'async'> {
     private finder: (specifier: string, options: CanonicalizeContext) => Promise<URL | null> | URL | null,
   ) {
     this.relativeImporter = new RelativeUrlRebasingImporter(entryDirectory, directoryCache);
+    this.canonicalize = this.canonicalize.bind(this);
+    this.load = this.load.bind(this);
   }
 
   async canonicalize(url: string, options: CanonicalizeContext): Promise<URL | null> {
@@ -370,17 +349,4 @@ export class LoadPathsUrlRebasingImporter extends RelativeUrlRebasingImporter {
 
     return result;
   }
-}
-
-/**
- * Workaround for Sass not calling instance methods with `this`.
- * The `canonicalize` and `load` methods will be bound to the class instance.
- * @param importer A Sass importer to bind.
- * @returns The bound Sass importer.
- */
-export function sassBindWorkaround<T extends Importer>(importer: T): T {
-  importer.canonicalize = importer.canonicalize.bind(importer);
-  importer.load = importer.load.bind(importer);
-
-  return importer;
 }
